backend/requirements.txt
# Web Framework & Server
fastapi
uvicorn

# Data Collection & Processing
yfinance
pandas
numpy
pandas-ta

# Machine Learning & Deep Learning
tensorflow
scikit-learn
joblib

# Scheduling & Environment Variables
apscheduler
python-dotenv
